+++ core/src/main/java/com/bakdata/dedupe/similarity/SimilarityContext.java
@@ -51,7 +51,7 @@ public class SimilarityContext {
     /**
      * Calculates the similarity when any of the two values under comparison is null.
      */
-    public <T> float getSimilarityForNull(T left, T right, SimilarityContext context) {
+    public <T> double getSimilarityForNull(T left, T right, SimilarityContext context) {
         return similarityMeasureForNull.getSimilarity(left, right, context);
     }
 }
